fix: made API_KEY nullable
